feat: Add MapOutputTool and ReduceOperationTool for MapReduce workflow
- Implemented MapOutputTool for recording Map stage processing results, managing task statuses, and generating output files.
- Added functionality for structured data input and dynamic parameter generation.
- Introduced ReduceOperationTool to support file manipulation operations (get_lines, append, replace) in the root plan directory.
- Enhanced error handling and logging for better traceability during operations.
- Established shared state management for coordinating multiple agent instances.
