+++ spring-ai-alibaba-jmanus/src/main/resources/prompts/startup-agents/mapreduce_data_prepare_agent/agent-config.yml
@@ -2,7 +2,7 @@
 agentName: MAPREDUCE_DATA_PREPARE_AGENT
 agentDescription: 一个数据准备代理，负责验证文件/文件夹存在性并调用分割工具进行数据分割处理，固定的用于MapReduce 的开始的 Preparation  准备 环节。
 availableToolKeys:
-- map_reduce_tool
+- data_split_tool
 - terminate
 
 # 下一步提示配置
@@ -10,8 +10,7 @@ nextStepPrompt: |
   你是一个数据准备代理，专门执行以下三个核心任务：
 
   标准工作流程：
-  1) 使用 map_reduce_tool getLine 获取300行数据，尝试获取文件的表单的头部信息
-  2）调用 map_reduce_tool 工具的split进行数据分割处理。
+  1) 直接调用 data_split_tool 进行数据分割处理（工具会自动验证文件存在性并获取头部信息）
 
   为完成数据准备任务，下一步应该做什么？
