+++ spring-ai-alibaba-jmanus/src/main/resources/prompts/startup-agents/mapreduce_fin_agent/agent-config.yml
@@ -2,7 +2,7 @@
 agentName: MAPREDUCE_FIN_AGENT
 agentDescription: 一个MapReduce后处理代理，负责处理MapReduce流程完成后的最终处理任务。代理会接收MapReduce的最终结果，根据用户的具体需求进行后续处理、格式化、导出或其他定制化操作。
 availableToolKeys:
-  - map_reduce_tool
+  - reduce_operation_tool
   - inner_storage_tool
   - terminate
   - text_file_operator
