+++ spring-ai-alibaba-jmanus/src/main/resources/prompts/startup-agents/mapreduce_reduce_task_agent/agent-config.yml
@@ -2,7 +2,7 @@
 agentName: MAPREDUCE_REDUCE_TASK_AGENT
 agentDescription: 一个Reduce任务执行代理，负责处理MapReduce流程中的Reduce阶段任务。代理会自动接收多个Map任务的输出结果，执行数据汇总、合并或聚合操作，并生成最终的处理结果。
 availableToolKeys:
-  - map_reduce_tool
+  - reduce_operation_tool
   - terminate
 
 # 下一步提示配置
@@ -12,7 +12,7 @@ nextStepPrompt: |
   简化的工作流程：
   1) 自动接收一批Map任务的输出结果（已通过上下文参数提供，无需手动读取文件）
   2) 根据操作步骤要求对多个Map的结果进行汇总、合并、聚合或综合分析
-  3) 使用 map_reduce_tool 保存最终的Reduce处理结果（你应只使用get_lines, append或replace操作！不要使用record_map_output或split_data操作）
+  3) 使用 reduce_operation_tool 保存最终的Reduce处理结果（你应只使用get_lines, append或replace操作！不要使用record_map_output或split_data操作）
   4) 输出完成后调用Terminate工具，结束任务
 
   为完成Reduce任务，下一步应该做什么？
